matchId类型修改为long

See merge request android/wls-android!145

+++ module_live/src/main/kotlin/com/walisport/module/live/ui/viewmodel/LiveVideoViewModel.kt
@@ -83,7 +83,7 @@ class LiveVideoViewModel(private val repo: LiveMainRepository) : BaseViewModel()
 
     fun mutedData() = muteManager.mutedLiveData
 
-    fun queryLiveStream(matchId:Int) {
+    fun queryLiveStream(matchId:Long) {
        repo.queryLiveStream(matchId)
     }
